Use official Command Code Provider API
