+++ index.ts
@@ -1,10 +1,11 @@
 /**
  * Command Code provider for pi.
  *
- * Connects pi to Command Code's API (https://api.commandcode.ai/alpha/generate).
+ * Uses the official Command Code Provider API:
+ *   https://api.commandcode.ai/provider/v1
  *
  * Authentication (pick one):
- *   1. Run `/login`, then select Command Code — opens browser to commandcode.ai, auto-stores API key
+ *   1. Run `/login`, then choose browser login or paste a Command Code API key
  *   2. Set COMMANDCODE_API_KEY environment variable
  *   3. Place API key in `~/.commandcode/auth.json` or `~/.pi/agent/auth.json`
  *      as {"apiKey": "user_..."} or {"commandcode": "user_..."}
@@ -12,21 +13,19 @@
  * Models are fetched from Command Code's Provider API at startup.
  */
 
-import { calculateCost, createAssistantMessageEventStream } from "@mariozechner/pi-ai"
 import type { ExtensionAPI } from "@mariozechner/pi-coding-agent"
 
-import { createStreamCommandCode, DEFAULT_API_BASE } from "./src/core.ts"
-import { DEFAULT_MODELS_URL, fetchCommandCodeModels } from "./src/models.ts"
+import { getConfiguredApiKey } from "./src/api-key.ts"
+import {
+  DEFAULT_MODELS_URL,
+  DEFAULT_PROVIDER_API_BASE,
+  baseUrlForModel,
+  fetchCommandCodeModels,
+} from "./src/models.ts"
 import { getApiKey, login, refreshToken } from "./src/oauth.ts"
 
-const API_BASE = process.env.COMMANDCODE_API_BASE ?? DEFAULT_API_BASE
+const API_BASE = process.env.COMMANDCODE_API_BASE ?? DEFAULT_PROVIDER_API_BASE
 const MODELS_URL = process.env.COMMANDCODE_MODELS_URL ?? DEFAULT_MODELS_URL
-
-const streamCommandCode = createStreamCommandCode({
-  createStream: createAssistantMessageEventStream,
-  calculateCost,
-  apiBase: API_BASE,
-})
 
 // ---------------------------------------------------------------------------
 // Extension entry point
@@ -38,14 +37,8 @@ export default async function (pi: ExtensionAPI) {
   pi.registerProvider("commandcode", {
     name: "Command Code",
     baseUrl: API_BASE,
-    apiKey: "COMMANDCODE_API_KEY",
-    authHeader: true,
-    api: "commandcode-custom",
-    streamSimple: streamCommandCode,
-    headers: {
-      "x-command-code-version": "0.24.1",
-      "x-cli-environment": "production",
-    },
+    apiKey: getConfiguredApiKey() ?? "COMMANDCODE_API_KEY",
+    api: "openai-completions",
     oauth: {
       name: "Command Code",
       login,
@@ -55,8 +48,10 @@ export default async function (pi: ExtensionAPI) {
     models: models.map((model) => ({
       id: model.id,
       name: model.name,
+      api: model.api,
+      baseUrl: baseUrlForModel(API_BASE, model.api),
       reasoning: model.reasoning,
-      input: ["text"] as const,
+      input: ["text"],
       cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
       contextWindow: model.contextWindow,
       maxTokens: model.maxTokens,
